refactor(collisions): extract broadphase to plugin
