+++ engine/CMakeLists.txt
@@ -47,8 +47,9 @@ set(CUBOS_ENGINE_SOURCE
     "src/cubos/engine/voxels/palette.cpp"
 
     "src/cubos/engine/collisions/plugin.cpp"
-    "src/cubos/engine/collisions/broad_phase.cpp"
-    "src/cubos/engine/collisions/broad_phase_collisions.cpp"
+    "src/cubos/engine/collisions/broad_phase/plugin.cpp"
+    "src/cubos/engine/collisions/broad_phase/sweep_and_prune.cpp"
+    "src/cubos/engine/collisions/broad_phase/candidates.cpp"
 
     "src/cubos/engine/input/plugin.cpp"
     "src/cubos/engine/input/input.cpp"
